pyproject.toml
[project]
name = "spark-device-library"
version = "0.1.0"
description = "Device library management for the ENEROOO Spark platform"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "HARDWARIO", email = "info@hardwario.com"},
]

dependencies = [
    # Django
    "django>=5.1",
    "django-environ>=0.12",
    "django-model-utils>=5.0",
    # API
    "djangorestframework>=3.14",
    "django-filter>=23.5",
    "drf-spectacular>=0.28",
    # Database
    "psycopg[binary]>=3.2",
    # Utilities
    "python-slugify>=8.0",
    "pyyaml>=6.0",
    # Static files
    "whitenoise>=6.7",
    # Production server
    "gunicorn>=22.0",
]

[project.optional-dependencies]
dev = [
    "spark-device-library",
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.10",
    "django-stubs>=5.0",
    "djangorestframework-stubs>=3.14",
    "pre-commit>=3.7",
    "django-debug-toolbar>=4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/config", "src/core", "src/library"]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["config", "core", "library"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "-v --reuse-db"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "environ",
    "debug_toolbar.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.local"
